Gebruikersbeheer, uitnodigingen en de volledige API-testsuite
Uitnodigen met een rol, rollen wijzigen en toegang intrekken, met een
uitnodigingstoken dat alleen als hash wordt bewaard. De policy voor
lidmaatschappen laat precies genoeg toe om een uitnodiging te accepteren:
alleen wie het token heeft, ziet de bijbehorende rij.

114 tests, alle groen: het volledige scenario, tenantisolatie via API en
rechtstreeks op de database met een metatest die elke tabel op een policy
controleert, rollen en rechten per handeling, aanmelden en MFA inclusief
TOTP-testvectoren uit RFC 6238, de bankbestandslezers, en de randgevallen
(onveranderbaarheid, creditnota, deelbetaling, betalingsverschil,
btw-verlegd, IC-levering, gesloten perioden, dubbele documenten, foutieve
imports, nummerreeksen onder gelijktijdigheid en vreemde valuta).

Onderweg gevonden en verholpen: een ORDER BY die op de tekst-alias sorteerde
waardoor de hash-ketting bij tien of meer regels leek te breken, jsonb die
sleutels herordent (nu canonieke JSON voor de hash), een tegenboeking die
zichzelf achteraf probeerde bij te werken terwijl definitieve posten
onveranderbaar zijn, een lege bedragkolom die stilzwijgend nul werd, en een
MT940-rekeningregel waar de valutacode aan het IBAN vastzat.

+++ apps/api/src/auth/service.ts
@@ -137,8 +137,10 @@ export async function meldAan(invoer: {
     await inTransactie(SYSTEEM_CONTEXT, async (client) => {
       const pogingen = gebruiker.mislukte_pogingen + 1;
       await client.query(
-        `UPDATE app_user SET mislukte_pogingen = $2,
-                geblokkeerd_tot = CASE WHEN $2 >= $3 THEN now() + interval '15 minutes' ELSE geblokkeerd_tot END
+        `UPDATE app_user SET mislukte_pogingen = $2::int,
+                geblokkeerd_tot = CASE WHEN $2::int >= $3::int
+                                       THEN now() + interval '15 minutes'
+                                       ELSE geblokkeerd_tot END
           WHERE id = $1`,
         [gebruiker.id, pogingen, MAX_MISLUKT],
       );
